linux-gen: arch: single cpu cycles file
CPU cycle functions were mixed over odp_cpu_cycles.c and
odp_cpu_arch.c files. Implement all three cpu cycle API
functions in odp_cpu_cycles.c. odp_cpu_arch.c is deleted.
