+++ platform/linux-generic/arch/x86/odp_cpu_cycles.c
@@ -4,14 +4,7 @@
  * SPDX-License-Identifier:     BSD-3-Clause
  */
 
-#include "config.h"
-
-#include <odp_posix_extensions.h>
-
 #include <odp/api/cpu.h>
-#include <odp_debug_internal.h>
-
-#include <time.h>
 
 uint64_t odp_cpu_cycles(void)
 {
@@ -28,4 +21,14 @@ uint64_t odp_cpu_cycles(void)
 		     "=d" (tsc.hi_32) : : "memory");
 
 	return tsc.tsc_64;
+}
+
+uint64_t odp_cpu_cycles_max(void)
+{
+	return UINT64_MAX;
+}
+
+uint64_t odp_cpu_cycles_resolution(void)
+{
+	return 1;
 }
